0x00-hello_world

+++ 0x00-hello_world/6-size.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
+#include <stdlib.h>
   /**
-    * main - A program that prints the size of varius computer types
+    * main - A program that prints the size of various computer types
     * Return : 0  (Success)
     */
 int main(void)
@@ -11,11 +12,11 @@ long int c;
 long long int d;
 float f;
 
-printf("Size of a char : %lu byte(s)\n", (unsigned long)sizeof(a));
-printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
-printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+printf("Size of a char : %lu byte(s)\n", (unsigned short)sizeof(a));
+printf("Size of an int: %lu byte(s)\n", (unsigned short)sizeof(b));
+printf("Size of a long int: %lu byte(s)\n", (unsigned short)sizeof(c));
+printf("Size of a long long int: %lu byte(s)\n", (unsigned short)sizeof(d));
+printf("Size of a float: %lu byte(s)\n", (unsigned short)sizeof(f));
 
 return (0);
